+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-09T13:33:10.011Z",
+    "2026-09-09T14:45:33.322Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {